nbody: use int instead of unsigned int for bodies

+++ nbody/src/nbody_lua_types/nbody_lua_type_marshal.c
@@ -68,7 +68,7 @@ static int readBodyArray(lua_State* luaSt, int table, Body* bodies, int n)
 }
 
 /* Read returned table of model components. Pops the n arguments */
-Body* readModels(lua_State* luaSt, int nModels, unsigned int* nOut)
+Body* readModels(lua_State* luaSt, int nModels, int* nOut)
 {
     int i, n, totalN, top;
     Body* allBodies;
@@ -104,7 +104,7 @@ Body* readModels(lua_State* luaSt, int nModels, unsigned int* nOut)
     lua_pop(luaSt, nModels - i);  /* No more body tables if error */
 
     if (nOut)
-        *nOut = (unsigned int) totalN;
+        *nOut = totalN;
 
     return allBodies;
 }
